test: extract word-list search tests into their own file
Move all search-related WordListContent tests (query filtering, the
tune icon, and the search-scope dialog) out of WordListScreenTest into
a dedicated WordListSearchTest, removing the need for the LargeClass
suppression the added tests introduced.
